refactor(gpu): 16x16 DSP multiply per lane (was 32x32 LUT) + synth notes
OOC synthesis showed the use_dsp="no" 32x32 LUT multiply (x8 lanes) was
impractical: ~16 min synth, ~16k LUT. Switch to a 16x16->32 signed multiply
mapped to one DSP48E1 per lane -> 8 DSP total (fits the ~18 free after the
NPU); "Finished Synthesize" drops to ~24 s. C golden model updated to match
(16-bit operands; exact for the demo kernels -- ALL GOLDEN CHECKS PASS).

docs/GPU_DESIGN.md gains a synthesis-notes section, also flagging the
async-read scratchpad inferring 65536 FF -> backing the scratchpad/imem with
BRAM (sync read) + mem_stall pipelining is the next implementation refinement.
Sim correctness (xsim, all 3 kernels) is unchanged.

+++ sim/gpu/gpu_model.c
@@ -33,8 +33,8 @@ static int32_t alu(int op,int32_t a,int32_t b,int32_t c){
       case VSRL:return (int32_t)(ua>>sh);
       case VSRA:return a>>sh;
       case VMIN:return a<b?a:b; case VMAX:return a>b?a:b;
-      case VMUL:return (int32_t)((int64_t)a*b);
-      case VMAC:return (int32_t)((int64_t)a*b + c);
+      case VMUL:return (int32_t)((int16_t)a*(int16_t)b);       /* 16x16 -> 32 (1 DSP/lane) */
+      case VMAC:return (int32_t)((int16_t)a*(int16_t)b + c);
     }
     return 0;
 }
